backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
bcrypt
python-jose[cryptography]
python-multipart
pydantic-settings
alembic
google-generativeai
